antitheft
Integrated anti theft security system for vehicle batteries. In the upcoming technology EV plays a key role in the vehicle manufacturing. In that the battery is important thing if it was theft means total vehicle will not move. In the cars the battery cost around 3- 4 lakhs . So for that we made a solution called Theftbocker.